+++ source/audiolib/src/multivoc.cpp
@@ -79,7 +79,7 @@ static VoiceNode  VoicePool;
 static int MV_MixPage;
 
 void (*MV_Printf)(const char *fmt, ...) = initprintf;
-static void (*MV_CallBackFunc)(uint32_t);
+static void (*MV_CallBackFunc)(intptr_t);
 
 char *MV_MixDestination;
 int MV_SampleSize = 1;
@@ -775,7 +775,7 @@ void MV_SetVolume(int volume)
 
 int MV_GetVolume(void) { return MV_TotalVolume; }
 
-void MV_SetCallBack(void (*function)(uint32_t)) { MV_CallBackFunc = function; }
+void MV_SetCallBack(void (*function)(intptr_t)) { MV_CallBackFunc = function; }
 
 void MV_SetReverseStereo(int setting) { MV_ReverseStereo = setting; }
